fix function

+++ Services/FavoriteService.cs
@@ -35,11 +35,11 @@ namespace Services
                 throw new Exception(e.Message);
             }
         }
-        public Task<bool> Add(Favorite favorite)
+        public async Task<bool> Add(Favorite favorite)
         {
             try
             {
-                return repository.Add(favorite);
+                return await repository.Add(favorite);
             }
             catch (Exception e)
             {
@@ -47,12 +47,20 @@ namespace Services
             }
         }
 
-        public Task<bool> Delete(string cusId, string recipeId)
+        public async Task<bool> Delete(string cusId, string recipeId)
         {
             try
             {
-                var fav = repository.GetAll(x => x.CustomerId == cusId && x.RecipeId == recipeId);
-                return repository.Delete(fav);
+                var fav = repository.GetAll(x => x.CustomerId == cusId);
+                var check = false;
+                foreach(var item in fav)
+                {
+                    if(item.RecipeId == recipeId)
+                    {
+                        check =  await repository.Delete(fav);
+                    }
+                }
+                return check;
             }
             catch (Exception e)
             {
